refactor: 애너테이션 기반으로 컨트롤러 클래스 재구조화
