+++ app/src/main/java/com/techcourse/controller/AuthController.java
@@ -12,12 +12,22 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Controller
-public class LoginController {
+public class AuthController {
 
-    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
+    private static final Logger log = LoggerFactory.getLogger(AuthController.class);
+
+    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
+    public String loginView(final HttpServletRequest req, final HttpServletResponse res) {
+        return UserSession.getUserFrom(req.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return "redirect:/index.jsp";
+                })
+                .orElse("/login.jsp");
+    }
 
     @RequestMapping(value = "/login", method = RequestMethod.POST)
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) {
+    public String login(final HttpServletRequest req, final HttpServletResponse res) {
         if (UserSession.isLoggedIn(req.getSession())) {
             return "redirect:/index.jsp";
         }
@@ -43,5 +53,12 @@ public class LoginController {
         } else {
             return "redirect:/401.jsp";
         }
+    }
+
+    @RequestMapping(value = "/logout", method = RequestMethod.GET)
+    public String logout(final HttpServletRequest req, final HttpServletResponse res) {
+        final var session = req.getSession();
+        session.removeAttribute(UserSession.SESSION_KEY);
+        return "redirect:/";
     }
 }
