--- app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,25 +1,0 @@
-package com.techcourse.controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.support.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Controller
-public class LoginViewController {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginViewController.class);
-
-    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) {
-        return UserSession.getUserFrom(req.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return "redirect:/index.jsp";
-                })
-                .orElse("/login.jsp");
-    }
-}
